requirements.txt
configparser
requests
solana
utils
pysolana
solders
reportlab
pillow
flask
